chore: use hooks
